refactor: organized file structure so that I can proceed with creating the rest of the pages
