+++ bootcamp-milestone-2/src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next"; // Object to set metadata
 import { Inter } from "next/font/google"; // You can change the font to anything you want.
 import localFont from "next/font/local";
-import "./globals.css";
+import "@styles/globals.css";
 import Navbar from "@/components/navbar";
 import Footer from "@/components/footer";
