+++ bootcamp-milestone-2/src/components/navbar.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import style from "./navbar.module.css";
+import style from "@styles/navbar.module.css";
 import Link from "next/link";
 
 export default function Navbar() {
