+++ bootcamp-milestone-2/src/typings/blog.ts
@@ -1,0 +1,8 @@
+export type Blog = {
+	name: string;
+	posted: string;
+	description: string;
+    image: string;
+    imageAlt: string;
+    slug: string;
+};
